requirements.txt
csv
sys
time
os
request
colorama
pyfiglet